ADM-store move board to config folder
