+++ frontend/__tests__/src/context/jiraVerifyResponseSlice.test.tsx
@@ -1,6 +1,6 @@
 import jiraVerifyResponseReducer, {
   updateJiraVerifyResponse,
-} from '@src/context/board/jiraVerifyResponse/jiraVerifyResponseSlice'
+} from '@src/context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 
 describe('jiraVerifyResponse reducer', () => {
   it('should show empty array when handle initial state', () => {
