+++ frontend/__tests__/src/utils/setupStoreUtil.tsx
@@ -1,9 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit'
 import { stepperSlice } from '@src/context/stepper/StepperSlice'
 import { configSlice } from '@src/context/config/configSlice'
-import { boardSlice } from '@src/context/board/boardSlice'
+import { boardSlice } from '@src/context/config/board/boardSlice'
 import { pipelineToolSlice } from '@src/context/pipelineTool/pipelineToolSlice'
-import { jiraVerifyResponseSlice } from '@src/context/board/jiraVerifyResponse/jiraVerifyResponseSlice'
+import { jiraVerifyResponseSlice } from '@src/context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 import { sourceControlSlice } from '@src/context/sourceControl/sourceControlSlice'
 import { metricsSlice } from '@src/context/Metrics/metricsSlice'
 import { pipelineMetricsSettingsSlice } from '@src/context/pipelineMetricsSettings/pipelineMetricsSettingsSlice'
