+++ frontend/src/components/Metrics/ConfigStep/Board/index.tsx
@@ -21,7 +21,7 @@ import { useVerifyBoardEffect } from '@src/hooks/useVerifyBoardEffect'
 import { ErrorNotification } from '@src/components/ErrorNotification'
 import { NoDoneCardPop } from '@src/components/Metrics/ConfigStep/NoDoneCardPop'
 import { Loading } from '@src/components/Loading'
-import { updateJiraVerifyResponse } from '@src/context/board/jiraVerifyResponse/jiraVerifyResponseSlice'
+import { updateJiraVerifyResponse } from '@src/context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 import { ResetButton, VerifyButton } from '@src/components/Common/Buttons'
 
 export const Board = () => {
