+++ frontend/src/context/config/configSlice.tsx
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit'
 import type { RootState } from '@src/store'
 import { PIPELINE_TOOL_TYPES, REGULAR_CALENDAR, SOURCE_CONTROL_TYPES } from '@src/constants'
-import { boardSlice as boardReducer, initialBoardState } from '@src/context/board/boardSlice'
+import { boardSlice as boardReducer, initialBoardState } from '@src/context/config/board/boardSlice'
 
 export interface configState {
   projectName: string
@@ -15,7 +15,7 @@ export interface configState {
   sourceControlFields: { sourceControl: string; token: string }
 }
 
-const initialState: configState = {
+const initialBasicConfigState: configState = {
   projectName: '',
   calendarType: REGULAR_CALENDAR,
   dateRange: {
@@ -37,7 +37,7 @@ export const configSlice = createSlice({
   name: 'config',
   initialState: {
     ...initialBoardState,
-    ...initialState,
+    ...initialBasicConfigState,
   },
   reducers: {
     updateProjectName: (state, action) => {
