+++ frontend/src/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit'
 import stepperReducer from './context/stepper/StepperSlice'
 import configReducer from './context/config/configSlice'
-import jiraVerifyResponseReducer from './context/board/jiraVerifyResponse/jiraVerifyResponseSlice'
+import jiraVerifyResponseReducer from './context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 import pipelineReducer from './context/pipelineTool/pipelineToolSlice'
 import pipelineToolResponseReducer from './context/pipelineTool/pipelineToolVerifyResponse/pipelineToolVerifyResponseSlice'
 import sourceControlReducer from './context/sourceControl/sourceControlSlice'
